src/requirements.txt
requests
python_dateutil
python_twitter
beautifulsoup4
aws-sam-cli